requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.8.0
black==25.1.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
fastapi==0.115.11
ffmpy==0.5.0
filelock==3.17.0
flake8==7.1.2
Flask==3.1.0
fsspec==2025.2.0
gradio==5.20.0
gradio_client==1.7.2
groovy==0.1.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.4.2
numpy==2.2.3
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
pyflakes==3.2.0
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
ruff==0.9.9
safehttpx==0.1.6
safetensors==0.5.3
semantic-version==2.10.0
setuptools==75.8.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.0
sympy==1.13.1
tokenizers==0.21.0
tomlkit==0.13.2
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.49.0
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
websockets==15.0
Werkzeug==3.1.3
